+++ doc/general/release-notes.rst
@@ -185,6 +185,7 @@ Other changes
 - The fields ``approximate`` (used to denote that a station's location
   has been assigned roughly) and ``asrid`` (altitude SRID) have been
   abolished. The field ``srid`` has been renamed to ``original_srid``.
+- The field ``Gentity.short_name`` has been renamed to ``Gentity.code``.
 - Station types have been abolished. Stations now don't have a type.
   The related information previously stored in the database will be
   deleted in the upgrade.